msglist: Parse content outside of build time
As impressive as it was that Flutter and Dart were giving us
perfectly usable levels of smooth scrolling in the rough initial
architecture here -- where every new message that came on screen
(or within a 250px window of it) meant parsing the message's HTML
and walking the resulting DOM tree, doing so again and again if
scrolling back and forth -- there was still some subtle jank
when scrolling rapidly through the list.

Fix that by applying the most basic piece of Flutter performance
advice: make the build methods fast by moving complex computations
elsewhere, doing them asynchronously.  Specifically, do the parsing
as part of the fetch step.