content: Fix breakage on headings above h6
This was introduced in the recent change that separated parsing
from building.  I added parsing of these headings (when previously
we'd treated them as unimplemented), but didn't add building for
them; and at the same time I made it an invariant that everything
we parse, we build.

This is one example of the kind of bug that the pattern-matching
feature in the upcoming Dart 3 would prevent!