msglist: Pad left and right insets
On some small devices in landscape mode, the `maxWidth: 760`
constraint doesn't (and isn't meant to) prevent the content from
extending into the left and right insets. So, use a SafeArea widget
for that.

In contrast to the SafeAreaView component from
react-native-safe-area-context in the RN app, this SafeArea widget:
  - knows when an ancestor has declared the given inset(s) consumed
    (by reading the ambient MediaQueryData.padding), and
  - declares the given inset(s) consumed (by changing its
    descendants' ambient MediaQueryData.padding).

It's left unaware of non-ancestors that consume insets. It turns out
that some insets are consumed by non-ancestors in this case; study
how that's done and add a few lines to adapt.

See the doc:
  https://api.flutter.dev/flutter/widgets/SafeArea-class.html

The `minimum: 8` should preserve the existing 8px padding in all the
places we want it, including:
  - phones with no notches/etc.
  - notched phones in portrait mode where the norm is to have no
    notches/etc. on the left or right.