ios build: Configure "Team" in Xcode's "Signing & Capabilities"
Without this, `flutter run` for iOS can give a build error that
says to go try building from Xcode, which in turn says to add this.

The team is "66KHCWMEYB", a.k.a. "Kandra Labs, Inc." (The Kandra
name appears in the Xcode GUI for me, but not in the committed Xcode
project config file, project.pbxproj.)

Making this change in the Xcode GUI added a line in three places in
the Xcode project config, one for each build configuration:
"Release", "Debug", and "Profile". Commit those lines.

The same line appears in the React Native app for its build
configurations, which are just "Release" and "Debug":
  https://github.com/zulip/zulip-mobile/blob/v27.200/ios/ZulipMobile.xcodeproj/project.pbxproj#L443
  https://github.com/zulip/zulip-mobile/blob/v27.200/ios/ZulipMobile.xcodeproj/project.pbxproj#L409