msglist: Fix double-padding of bottom inset
The user-visible effect was pretty small, so I didn't debug it fully
in PR #3: on my iPhone, before this commit, scrolling all the way
down in the message list left about a centimeter, maybe less,
between the bottom of the latest message and the top of the
compose-box placeholder.

Now, on my iPhone, scrolling down puts the bottom edge of the latest
message flush with the top of the compose-box placeholder, as has
been the case on devices without insets. I'm not sure if that's
quite the effect we want, but if we want to add padding, we should
do it on purpose.

Addressing Greg's TODO in 88f3b73a6 would have prevented this bug
from sneaking in; I'll take care of that next:

  // TODO have MessageListPage adjust the MediaQuery so we don't
  //   have to worry about this here