android: Optionally sign with a release key
This is much the same setup as we've had in the React Native app:
by default, release mode still uses the debug signing key, but
you can pass `-Psigned` to use a release signing key instead.

That flag applies to the `flutter` CLI tool whenever building
the Android app: `flutter run`, `flutter build appbundle`, or
`flutter build apk`, any of which will pass it on to Gradle.