store [nfc]: Split GlobalStore for easier testing
This lets a test subclass GlobalStore to provide a different
way to load a PerAccountStore.

(Widgets still hardcode the "live" GlobalStore, so we'll need
to make some further changes when we go to write tests for those
widgets.  But this is already enough to write tests for any
logic we put on GlobalStore.)