store [nfc]: Fold processInitialSnapshot into a constructor
This is how it's really functioning anyway.

This arrangement will be helpful as we split PerAccountStore into
a base class and an implementation with more details, as the latter
can then have a constructor that cleanly delegates most of the work
to the former's constructor.

This way also reduces the amount of boilerplate required to add
another piece of data to the model.