widgets/app [nfc]: Use two-space indents consistently
Much of this code was written before I had a firm grasp on how
indentation in Dart code is supposed to work.  Among other things,
this meant I often let the auto-formatting in Android Studio do
what it seemed to want, even when it was confused.

This diff should be empty under `git diff -b`.  I'll make
small other tweaks as a separate commit.