login: Add a basic login flow
There's plenty more to do to improve the UX of this, and I've
filed several follow-up issues.  Most critically, we'll need
to store the resulting credentials on the device, so you don't
have to repeat the login each time you start the app.

But this is enough that it becomes possible to log in at an
account of your choice, making the flow testable end to end.
In particular this makes a useful bootstrapping step toward
the task of storing data locally: with this, there's now some
useful data we could store.

Fixes: #11