api [nfc]: Use Dart-standard camel case throughout our source code
If the field names and parameter names used in the API bindings
match the snake case of the Zulip API, then that will unavoidably
propagate to some names found throughout the rest of our code --
as indeed it already has, as seen in this diff.  There's no clean line
that can be drawn within our source code so that snake-case names
go on one side, and camel-case names on the other.

So we'd inevitably have an unpredictable mix of both kinds of names,
and often of the same name appearing in two different forms in the
same code.  That's what we see in the zulip-mobile RN app.

Instead, draw that line at the very boundary of our code, where
the data comes and goes via JSON.  Happily the json_serializable
code generator allows us to largely just write camel case and let
the conversion to and from snake case be done automatically.

Written by removing those ignore_for_file lines, and then using
Android Studio's "rename" feature many times.  Then grepped for
`\w+(_\w+)+` to find stragglers in comments, and fixed those by hand.

There's a couple of remaining stragglers in credential_fixture.dart,
which I'll handle as a followup commit.