compose [nfc]: Rename ContentTextEditingController's listener
TopicTextEditingController and ContentTextEditingController inherit
from ValueNotifier, so they have a `value` property, and their
listeners are notified when that property changes. They don't
currently notify their listeners in any other cases, so the
listeners are reasonable to have "value" in their names.

But for ContentTextEditingController, soon we'll want to call
notifyListeners for something other than a change in `value`: in
particular, it'll have a stateful property to track in-progress file
uploads, and that property will affect `validationErrors`, so we'll
call `notifyListeners` when a file upload starts or ends. So, we
give its listener this more general name.