login: Use canonical realm URL from server
Just the other day we were talking about this get-server-settings
property, and I explained why it's important for a mobile client
to use it:
  https://chat.zulip.org/#narrow/stream/378-api-design/topic/uri.20to.20url.20compatibility/near/1549242

But apparently when I wrote this code a couple of weeks earlier,
I'd forgotten about that, oops.  Fix it.

It's particularly useful that we fix this before we start keeping
the list of accounts in a database, so that there's no legacy of
some people's devices having a non-canonicalized URL lying around.