login: Parse and validate realm-url input more like we do in zulip-mobile
For the corresponding logic in zulip-mobile, see
src/start/RealmInputScreen.js.

Between the URL parser we use here and the one we use in
zulip-mobile, probably the set of rejected strings won't match
exactly, but it'd be a surprising bug if either one rejects
reasonable strings that we could've used.

As in zulip-mobile, we distinguish validation errors we want to
report immediately, from ones we'll wait to report until the user
presses submit. (Compare this revision's `shouldDeferFeedback` to
the first bit of zulip-mobile's `getSuggestion`.)

Also add some tests, in an improvement over zulip-mobile.

Fixes-partly: #35