login: In username field, give "next" button instead of "done"
This does a nice thing beyond just using more appropriate text for
the button: when the button is pressed, focus automatically moves to
the password field.

Too bad it doesn't call the email field's `validator` before
deciding to change focus, as noted in the comment. We could probably
implement that ourselves, but I suspect that'd require managing a
FocusNode for the email field, which is just an extra bit of
complexity.