fake_api: Use auth in FakeApiConnection.fromAccount constructor
Seems most logical, given that the account contains the auth.
And this will enable us to test that the auth information gets
appropriately included in the HTTP request.

(The API key should of course be a fake one; the tests shouldn't
have any other kind of API key in the first place.  And the
fake connection won't generate any real network requests in
any case, because it uses a FakeHttpClient.)