login [nfc]: Use "username/password" variable names, not "email/password"
Unless serverSettings.requireEmailFormatUsernames is true, the
"username" part of username/password login might not be an email
address. Change our variable names and a comment to use the more
general term "username", which can apply whether or not the string
is an email address. (`username` is in fact the name of the
fetchApiKey param.)

For what we were calling EmailPasswordLoginPage, just name it
PasswordLoginPage. As Greg points out:
  https://github.com/zulip/zulip-flutter/pull/91#discussion_r1194292191

> In general, the phrase "password login" is more common for this
> kind of login flow than "username/password login". And in the
> Zulip API (https://zulip.com/api/get-server-settings), we call
> this auth method simply `password`.

Next, we'll make functional changes to explicitly support both
possible values of serverSettings.requireEmailFormatUsernames.