deps: Sync ios/Podfile.lock with recent dependency upgrades
These changes get made automatically for me when I build/run the iOS
app, which I believe is when `pod install` gets run.

It looks like a symptom of `pod install` not being done in 861a92911
(in the case of `path_provider_foundation`) and in 99dd6a7e4 (in the
cases of `device_info_plus` and `share_plus`).