content: Have `errorCodeStyle` ask for "Source Code Pro"
Like in the other places we ask for a monospace font.

(The exact choice of font isn't actually important here; the text is
just for displaying what message-content elements we haven't
implemented yet. But we're about to define a TextStyle constant to
encapsulate reasonable font-family fallbacks for a monospace font,
and it'll be convenient to use that constant here too. It'll use
"Source Code Pro" as its first choice.)