text: Set up to use variable fonts with a "wght" axis
For background on our interest in variable fonts with a "wght" axis,
see #65.

One thing I noticed when adding such a font ('Source Code Pro', in a
commit coming soon) is that by default, the text was drawn really
lightly; in fact, Flutter seemed to be using the font's lightest
weight by default. I guess I'd assumed Flutter would pick a "normal"
weight by default, as it does for non-variable fonts. Possibly
that's just because Flutter hasn't fully caught up with this recent
development in fonts. But also, I'm not sure if these new fonts all
*have* mappings from "normal", "bold", etc., to values on their
"wght" axes...whether such mappings are standard/predictable, or at
least declared by the font in structured metadata.

Anyway, in my design here, it means that if you want to use one of
these fonts, you still need an incantation (weightVariableTextStyle)
to render text where a normal weight makes sense. Hopefully that's
not too burdensome, but it comes with a benefit: as long as that's
the norm, I think we're unlikely to misuse `fontWeight` to
control glyphs that want to be controlled by "wght".

Much of the complexity in this commit comes from handling glyphs
that need to be rendered in a fallback font that doesn't have a
"wght" axis; see clampVariableFontWeight and how we use that. It
means those glyphs will have approximately the weight of the other
glyphs...not essential, but nice to have.

Some complexity comes from extending Flutter's support for the
"bold-text" accessibility setting, so that it applies to these new
fonts that use "wght". Ideally we wouldn't regress on our support
for that platform setting, but in this design, it means callers have
to have a BuildContext on hand; hmm.