sticky_header: Pass new second type argument to SlottedContainerRenderObjectMixin et al.
This is required since https://github.com/flutter/flutter/pull/126108 .

The breakage is a reminder that it'd be good to get our repo added
to Flutter's "customer tests" suite.  Because this isn't a common API
to extend, neither that suite nor Google's internal use of Flutter
had anything affected by this change, which mean that Flutter's
"breaking change" policy treated it as a non-breaking change, the
same as an internal refactoring.

In addition to these required changes, there's a deprecation
of SlottedMultiChildRenderObjectWidgetMixin in favor of an abstract
base class.  We can take care of that in a separate commit.

[greg: wrote new commit message]