fmt [nfc]: Adjust newline placement to Flutter style
From the Flutter repo I've picked up a preference to avoid a
line break just after `=` or `=>`, and in a few similar "cliffhanger"
locations.  This is paired with a willingness to make some lines
longer than 80 columns when that helps readability:
  https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo#prefer-a-maximum-line-length-of-80-characters
  https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo#consider-using-for-short-functions-and-methods

We have some such "cliffhanger" newlines in our existing code,
much of it in places where the Dart autoformatter put it there
before I'd gotten a grasp on what I want to keep from the
autoformatter's output and what I don't.  So here's a quick sweep
through our code fixing those.

Conversely, fix a couple of too-long lines, and mismatched formatting
of parallel constructs, that I noticed while doing the sweep.