nav: Add button for test-here narrow
Up to now we've provided navigation only to the all-messages narrow,
and we've given that screen a compose box with a hack that has it
send only to `#test here`.

We're going to remove that hack and put no compose box on that screen.
To prepare for that, we'll want some other way to quickly get to a
screen with a compose box, for testing.  Do that with the actual
stream narrow for `#test here`.