compose: Drop chat.zulip.org-specific scaffolding hack
Now that one can navigate to the actual stream narrow for the
`#test here` stream on chat.zulip.org, there's no longer a need
for the hack that offered that stream's compose box on the
all-messages narrow's screen.

And then with that hack gone, the send button works correctly
on any realm; so we can drop the condition that made it refuse
to operate outside of chat.zulip.org.