api [nfc]: Reorder and rearrange Subscription fields to parallel Stream
This type is long enough to kind of push the limit for manually
comparing the docs to the type when they're not in the same order.
But in practice for future updates to it I think we can rely on
reading through its API docs for "Changes" notes.

And meanwhile, even within the API docs the `subscriptions` and the
`streams` lists don't agree on the order of the fields, even among
those they have in common.  So we might as well give up on matching
them and go for a logical order instead.