compose: Add proper didUpdateWidget for notifier subscriptions
This is a required part of the protocol for a state node to
listen to objects that can change, as described in the docs for
[State.initState] and [State.dispose].

I don't think it's real likely to come up for these particular
widgets in practice -- I think they probably never will have
the controllers swapped out on an existing state node.  But
best to follow the protocol anyway, as part of a pattern of
systematically doing so.