test [nfc]: Clean up account/ID match in model/store_test
When this test file was originally written, an Account object didn't
contain the account ID; the ID lived only in the key in this map,
so it wasn't possible to have a mismatch between key and value.

Now an Account object contains its own ID, so to keep the data
consistent, the map should agree.  Fix up this test data accordingly.