store [nfc]: Drop redundant account IDs at GlobalStore constructor
When I first wrote down this constructor, an Account object didn't
contain its account ID, so the map was needed in order to express
those.  Now it does, so it's cleaner to just accept an iterable
and construct the map internally: a bit simpler for callers, plus
by eliminating the redundancy it eliminates a potential form of
inconsistency in the data.