content: Use "growable: false" optimization in _buildInlineList
Making lists non-growable is a small optimization that's
particularly helpful when creating a large number of short lists.

That describes what we do here, mapping inline content nodes to
corresponding inline spans: there will be tons of these nodes in a
message list; tons of them will be in lists of one element (like
any paragraph of plain text, or any link or strong/bold or
emphasis/italic span with no further markup inside it); and
probably the bulk of the rest will be in lists of two or three.

So this is a spot where the optimization may well have a material
performance impact.  In any case it's essentially free; since we
didn't want to add or remove elements anyway, the only cost is the
extra few tokens in the source code.