content [nfc]: Expand cryptic comment on UserMentionNode
I wasn't sure what this comment meant when I came across it recently,
even though I wrote it myself.  Today I was writing tests for this
code and saw what the data looks like, and the meaning became clear.
Expand the comment so that it's hopefully clear to other readers
without having to see the data.