content test: Compare parse trees via Diagnosticable
This makes for significantly cleaner and more informative output
when a test fails (resolving all the TODO-comments), by taking
more control of how the comparison is done.

At the same time, it makes for less boilerplate and fewer places
to update when adding more features in the ContentNode parse trees.
Effectively it deduplicates some of the boilerplate, by reusing
for comparing the trees the metadata that we're already prepared
to generate for printing them out when a comparison fails.