binding [nfc]: Rename ZulipBinding from DataBinding
We'll need some more bindings in this spirit.  It's not clear there'll
be enough of them that we need to split them into mixins across
multiple files, like Flutter upstream does, so we'll try just putting
them all in a single bindings class.  Start by giving that one class a
more generic name to fit.