narrow: Implement `toString` on {Stream,Topic,Dm}Narrow classes
This will be helpful for test-failure output, particularly with
DmNarrows as we start using those as Map keys, soon.

Of our current Narrow subclasses, this just leaves AllMessagesNarrow
without a custom `toString` implementation. That's OK, because the
default `toString` already says "AllMessagesNarrow", which is all we
need to know about the narrow.