test [nfc]: Add shorter name "testBinding" for TestZulipBinding.instance
We use this a lot in our tests, and its name is kind of uncomfortably
long for that.  Make it shorter.

In particular if in a test one wants the test binding object for
Flutter itself (the TestWidgetsFlutterBinding), that's spelled
`tester.binding` (where `tester` is the argument passed to the
`testWidgets` callback).  So this makes our own test binding object
be comparably accessible, as `testBinding`.