vscode: Disable reformatting whole swaths of code
The details here are largely borrowed from the Flutter repo,
and in particular the subtleties most recently added there:
  https://github.com/flutter/flutter/pull/122758

This config does the job in my testing.

Note that it doesn't affect any of the things the editor does to
simply augment your typing: adding indentation when you hit enter,
adding a close-paren when you type an open-parent, and so on.
Those are perfectly fine -- most of all because they only affect the
code you actually intended to edit, but also because they generally
get things right -- and continue operating as usual.

Fixes: #229