autocomplete [nfc]: Use new mixin for ComposeAutocomplete's state
As with other uses of PerAccountStoreAwareStateMixin, we expect the
benefit to be noticeable after our planned implementation of #185,
"Start new event queue when old one has expired".

At that time, this should mean that an expiring event queue won't
break autocomplete: if the lifetime of your autocomplete intent
spans across a queue-renewal, you can continue to edit your query
after the renewal, and the app will ask the new store for the query
results, instead of asking the old one.