api: Add user_settings/update events, with a test for exhaustiveness
It's been tricky to find a way to verify that the event-handling
code keeps up with the settings we add in [UserSettings], the data
class we use in the initial snapshot. See #261 for some alternatives
we considered.

But at least this solution works, with type-checking of the event at
the edge, and a mechanism to ensure that all user settings we store
in our initial snapshot get updated by the user_settings/update
event.