msglist test [nfc]: Make "before" checks be more to the point
This test is for `maybeUpdateMessage`, not `fetch`.  So it's
not its job to check that `fetch` successfully got the content
and flags and isMeMessage that we expected.

What is useful for a "before" check here, on the other hand,
is to confirm that our "after" checks aren't going to be vacuous,
by checking that the message doesn't already have the values
we'll be looking for there.  So check that instead.  Include
one for lastEditTimestamp, which was previously left out of
the "before".

A low-tech way to express this would be to keep the checks
like `.equals(oldContent)` and add assertions that oldContent
differed from newContent, and so on.  But we can do it a
bit more neatly using the `not` check.