msglist [nfc]: Set route settings on message-list page route, for testing
This way, the Route object itself is distinguishable from other
Route objects. This will come in handy in widget tests for the
upcoming recent-DMs screen (#119), where we'll want to assert that
tapping on a conversation brings up the appropriate message list.

While we're at it, add the usual
`extension FooChecks on Subject<Foo>` boilerplate that we'll use in
those tests.