msglist test [nfc]: Drop user and stream from store
The code under test doesn't actually look at this information --
the only thing it uses the PerAccountStore for is to (a) get an
API connection, and (b) unregister itself in `dispose`.
So simplify the tests by leaving it out.

If later the code under test does start needing this information
so that we do need to set it up in the tests, we can revisit then
and may find a bit simpler of a way to arrange it.