msglist test: Overhaul setup of tests, with shared helpers
All the tests in this file are going to have a common structure
to the objects they operate on: chiefly, a MessageListView.
So we can hoist those up to the level of `main`, which enables
the tests to share helper functions that act on those objects.

In principle this means a state leak between tests.  We mitigate
that by resetting all the shared variables at the same time,
in the `prepare` function that each test calls at the outset.

We could get a similar effect by putting these shared variables and
helpers as fields and methods on a class, and having each test case
make its own instance of the class -- its own "tester" object, akin to
the WidgetTester object that `testWidgets` provides to its test body.
That'd provide more encapsulation, and would be a good strategy if
for example this logic were to be shared across several test files.
But I think this test file can get away without it.

The immediate payoff of the shared helpers is that they let us
centralize the logic for checking that the view-model has, or hasn't,
notified its listeners.  That simplifies each test case, plus it
lets us make that logic slightly more sophisticated (at O(1) rather
than O(n) cost in the source code): we now detect duplicate calls too.

Coming up, we'll add further centralized useful logic to these helpers,
and add some tests that benefit from the split between `prepare`
and `prepareMessages`.