api: Implement toString on NetworkException
In particular, include the underlying cause.  As with
ZulipApiException, this makes the debug log more helpful when
one of these errors is encountered in development or in a test.

One UI test needs updating with this change, reflecting that we're
not consistent about how we extract from a failed API request a
message to show to the user.  Leave a comment at the relevant
spot, and use a `TODO` tag to cross-link it with an existing
TODO comment in some existing code we'll want to deduplicate.