api: Add unreadMsgs in initial snapshot
Our own data structure in PerAccountStore will be constructed from
this value, and it will probably have quite a different shape; for
example, there, we're likely to want to collapse the distinction
between 1:1 and group DMs.

But now we're validating this part of the /register response and
making it available to feed into those new internal data structures.

Related: #253