msglist: Show message reactions!
And also support:
- removing a reaction you've already made, and
- joining in on existing reactions that other people have made.

As is our habit with the message list, this aims to be faithful to
the web app, as accessed today. That should be a good baseline to
make mobile-specific adjustments from. (In particular I think we may
want larger touch targets.)

Unlike the web app, we use a font instead of a sprite sheet to
render Unicode emoji. This means that we, unlike web, have to
account for text-layout algorithms, and things like font metrics. So
if Unicode emoji appear noticeably differently from web, that
probably has something to do with it.

Fixes: #121
Fixes-partly: #125