model [nfc]: Implement ReactionType.toJson
We'll call this directly soon, when we'll want reusable code to
translate an enum value into a 'snake_case' string, for API bindings
for the add- and remove-reaction endpoints.

With a `toJson` method, instances can now be handled by
`encodeJson`. That's nice, but not really a goal here; we don't have
an imminent need to pass an instance through `encodeJson`. Anyway,
it means that json_serializable's generated _$ReactionToJson and
_$ReactionEventToJson functions don't need the enum-to-string
conversion on their reactionType fields in order for jsonEncode to
accept those functions' output. And it looks like json_serializable
has removed that conversion in both functions. So, adjust our code
to not depend on it.