deps: Upgrade Flutter to latest main, 3.14.0-5.0.pre.23
And update Flutter's supporting libraries to match.

Also remove a `!` operator that Dart 3.2 identifies as redundant.
(Happily, the change seems to be controlled by the version specified
in our pubspec.yaml, not by the version of the Dart binary -- so a
newer Dart binary will accept our code both before and after this
change.)