content [nfc]: Cut down TODOs on _launchUrl, following investigation
The second TODO, about the behavior of externalApplication, turns out
to be just how things work these days on Android:
  https://github.com/zulip/zulip-flutter/issues/74#issuecomment-1681672754

The first one, about the behavior of inAppWebView or platformDefault,
I've filed as #279.  I also added a number of details on the upstream
issue, and cross-linked it with a couple of other issues that would
be fixed by the same solution.