msglist test: Use default screen size
This size is in *physical* pixels, not logical -- so with the
tests' default device pixel ratio of 3, it corresponds to
200w267h in logical pixels, which is a very small screen.
(It's a bit over 2 inches (53mm) diagonal when converted at
160px/in, which is the nominal value for a logical pixel.)

Better to use the default size, which is somewhat more realistic.
In particular, we have a known issue (#282) that the recipient headers
suffer overflow if a stream name is long.  This tiny screen causes
that issue to appear even when a name is not very long at all;
and when we add dates to the recipient headers (#172), it would
start happening in the tests as they stand.

A handful of tests need adjustments to keep working with the new
size, mostly just to fill the message list with more messages.