msglist test: Make double-fetch-glitch test robust to layout changes
Since adding this test, we've repeatedly seen it break upon
changes that affect how tall the messages are:
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/flutter.3A.20double-fetch.20glitch.20gone.3F/near/1631114

In particular, the length of this pump for the fling-scroll has
needed adjustments.  Automate that by just finding out empirically
how far is far enough to trigger the fetch, and proceeding from there.