sticky_header test: Add tests for sticky headers
We're about to replace this library with a fresh implementation
that has a different design, in order to gain some features
we'll want for the message list.  These tests will help us
validate the new implementation.

Some of the naming in this test code reflects the interface the
new implementation will have, in order to changes in the tests
and so maximize their effectiveness in validating the reimplementation.