ui: For boring, non-emoji font, use one that iOS recognizes
It's strange that iOS really doesn't seem to know what to do with
'sans-serif' here. Nor "SF Pro", "SF-Pro", nor any obvious variation
on that, even though it's supposedly the system font:
  https://developer.apple.com/fonts/

But a GitHub comment popped up a few hours ago (linked in the added
code), and that's the first thing I've found that seems to work, as
a clear pointer to the default system font. Shrug.

Later, we'll want to use Source Sans 3 for almost all the text in the
app. But we'll want to take more care with that than I have
bandwidth for right now (around it being a variable-weight font);
see #294.

Fixes: #289