model: Use enum for Message.flags[]
It would have been nice to be able to annotate [MessageFlag.unknown]
as the "unknown" value inline, right there in the enum definition.

If the enum described a field directly on the Message class, we
could at least have annotated that field with
  @JsonKey(unknownEnumValue: MessageFlag.unknown)
and that would be nice and compact. But the field is a
List<MessageFlag>, not a MessageFlag, so we have to verbosely write
out `fromJson` for it, ah, well.