docs: Make an editing pass on new translation docs
Mostly this revises the text to be more fully oriented
toward giving instructions on what one needs to know
for working on the app, and in particular for adding
UI features.

Also add the fun and helpful fact that a hot reload is enough
to cause the bindings to get updated from the ARB files.

And tweak the descriptions on a couple of strings;
these label just a single item, not a section.