check: Add suite shellcheck
This is based on a similar suite I drafted for zulip-mobile
back in... 2019, yikes.

It looks like at the time the thing that stopped me from merging it
there was the question of how to handle Shellcheck as a dependency.
But that needn't be a barrier to merging it -- it just needs to go
into the list of "extra suites" that don't run by default.
As a bonus we can give a nice error message when missing.