api/notif: Parse notification messages the server sends over FCM
Much of this is based on the Kotlin code that does this same job
in the zulip-mobile RN app:
  android/app/src/main/java/com/zulipmobile/notifications/FcmMessage.kt
  android/app/src/test/java/com/zulipmobile/notifications/FcmMessageTest.kt

Those Kotlin tests benefit from a feature of their test framework
where when an expectation like `expect.that(foo).isEqualTo(bar)`
(analogous to our `check(foo).equals(bar)`) is unsatisfied, it
records the failure without immediately ending the test function.
As a result, when something goes wrong we get very fine-grained
results, much like we would with package:checks if all the test
functions were translated as `group` and the individual expectations
as `test`, rather than as `check` -- but without the overhead that
the latter approach requires in the source code, particularly in
making up names for the individual test cases.

In this version, we go ahead and translate to `group` and `test` for
a couple of the test functions where their lack would be most keenly
felt, resorting to a hack to avoid cluttering the code with lots of
individual names.  We let the rest be just `test` and `check`.